UnreadCards: Avoid collision between strings from users and a fixed string.
We'd like to clarify and untangle most of the UI code in
UnreadCards, StreamItem, and elsewhere (part of that is #3767), but
the most pressing task is to minimally fix an issue where a stream
named "private" causes bugs for other users who are subscribed to it
and are viewing UnreadCards on mobile.

So, fix it at the source, where the `key` in the data is first
assigned, by adding a prefix. Fortunately, consumers of this data
don't ever expect that the `key` will be equal to the stream's name.

The prefix wouldn't be necessary if we made this two separate lists in
the first place, which might be cleaner than the single SectionList;
for example, a FlatList for PMs, and a SectionList for the
stream/topic messages. PMs just don't fit in a heading–item structure,
like streams and topics do, much less one with the same semantics.