nav: At login, wait for load just like on account switch.
The app's main UI relies on actually having data from the server --
after all, without that there's not much useful it could do anyway --
so we should wait to show that UI until we have that data.

The visible effect of this is that on login, instead of the
broken-looking state where the lower-right corner where a tab icon
should be is missing/blank (because it's meant to be the user's
avatar, and we don't know their avatar URL yet) and we wrongly say
there are no unreads, we just show the loading screen until we're
ready to show the real data.

This is already what we do in the case where the user is switching
from one account to another.  There's no reason we should think we're
*more* ready to show the main UI when it's an account we weren't even
logged into until just now.

In fact the only reason this hasn't been a live crashing bug for a
long time is that we have in the realm reducer a hack to provide a
couple morsels of fake data in this case, causing much of the UI to
simply show nonsense (like that blank avatar) instead of crashing.
We'll get to take that hack out next.

Fixes: #4077