api [nfc]: Remove now-unused function `objectToParams`.
This function was effectively untyped; most misuses of it would not be
caught.

It could probably be improved and/or given a proper type... but, as it
turns out, all but one of this function's existing callsites were
unneeded, and the last was easy to inline cleanly.

So, send it to the cutting-room floor.