android build [nfc]: Remove residual comment from WebView being in RN core.
Part of the RN v0.59 -> v0.60 changes to the template app [1].
Technically, this should happen after the upgrade commit because RN
v0.60.0 newly omits WebView in RN core. But we're not using the
WebView in RN core, we're using `react-native-webview`, from
`react-native-community`. So the absence of this hint will be fine.

Corresponds to facebook/react-native@6345fcf12, released in RN
v0.60.0.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.6