Revert "sentry: update and fix embedded Xcode build scripts"
This reverts commit b4c11e02a, fixing #3809.

The effect of that commit was to change the conditions under which
our iOS build would upload debugging information to Sentry.  The
old logic was sloppy, but it worked correctly for real releases.
The new logic means we never successfully tell Sentry about an iOS
release -- we only try in conditions where we can never succeed.
As a result it becomes much harder to debug any failure specific
to iOS.

We'll file a followup issue to fix the old logic's sloppiness in a
way that works.  Until then, revert.

Fixes: #3809