ios deps [nfc]: Mark two deps that we're explicitly not linking.
In 33f4b41c8 (the CocoaPods commit), we sectioned off a "Pods we
need that depend on React Native" list. First, we accumulated these
dependencies one by one as we observed build failures caused by
their absence. (It's self-evident that we did this, at least, but
I'm fairly certain we also went down a list in the Xcode UI of
libraries that were linked.) Then, we checked this list in the
Podfile against a search of libraries in node_modules with a
"podspec" file that declared a dependency on React Native.

From that commit:

"""
We also added a number of dependencies that depend on React Native
(react-native-image-picker, etc.).
`grep -Rlw 'React' --include=\*.podspec node_modules/` verified
that they do indeed depend on the 'React' pod, which is React
Native.
"""

Now, while working on the RN v0.60.0 upgrade, I decided to re-check
that list with a more precise search:

```
grep -Rl dependency.*React --include=\*.podspec \
  --exclude="node_modules/react-native/*" node_modules
```

-- we were looking for non-RN libraries that have a podspec with a
line like `s.dependency 'React'`.

The results of this search are much cleaner. It

    (a) excludes podspec files in `react-native`, and

    (b) excludes offhand, code-comment mentions of React.

(b) didn't change the result set at all, but (a) reduced a lot of
noise. Following cb87f9062, there's a result from `@unimodules`
which we can ignore; that's handled by `use_unimodules!`.

I re-checked that everything in that list in our Podfile matched one
of these search results; so far, so good. But with this cleaner
result set, I had the idea to check the reverse: that each item in
the result set (except the irrelevant `@unimodules` one) was
represented in that list in the Podfile.

Two results were not: `react-native-document-picker` and
`react-native-screens`.

I verified that, in fact, we weren't linking either of these before
the CocoaPods commit. But it's best to mark this explicitly, with my
(new, alas) knowledge that we're excluding them intentionally.

Some background on why we didn't link them --

On iOS, there are currently no uses of
`react-native-document-picker`; it's only imported in
`src/compose/ComposeMenu.js`, and there, only to enable file uploads
on Android. See d9bbb5e5f, and the discussion at #3490, where it
looks like we explicitly decided *not* to link it on iOS -- some
stuff in `project.pbxproj` was auto-generated, as we'd expect before
CocoaPods, and Greg removed it as unnecessary.

Piecing together the `react-native-screens` story was more
interesting, and it resulted in filing #M4111. See also discussion
in chat [1]. We haven't really known what this library can do for
us, but now we sort of do. Anyway, we explicitly decided not to link
it during work on 0d2066f41; see discussion [2]. That will change
during work on #4111.

[1]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/iOS.3A.20r-n-screens.20.2F.20r-n-document-picker/near/877096
[2]: https://chat.zulip.org/#narrow/stream/48-mobile/topic/Windows.20build/near/794533