android build: Prepare to enable Hermes.
See #4131, the issue for enabling Hermes.

Part of the RN v0.59 -> v0.60 upgrade [1]. This must happen at or
after the upgrade commit because the use of Hermes is newly
supported.

RN did this in two main steps that affected the template app, which
we squash together in this commit.

RN's first step was to start using NPM to manage the JSC version;
this was released in v0.60.0:

- facebook/react-native@8e375850d

- facebook/react-native@4bb0b4f20 (a partial reversion)

RN's second step was a large amount of setup to get Hermes ready,
released in v0.60.1:

- facebook/react-native@e857d7066

- Then facebook/react-native@0738fe573 fixed a small bug, released
  in v0.60.2.

All four of the above-mentioned commits are reflected in this
commit.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.6