flow: Use `inexact-spread` lint rule.
Preparatory to the RN v0.59 -> v0.60 upgrade [1].

In facebook/react-native@0f909e331 (released in v0.60.0), the
`inexact-spread` rule is activated in the template app, at the
"warn" level. We want to make this change in our own app.

Unfortunately, this lint rule isn't documented. With that in mind,
we can address some of its output with suppression, where it proves
at all complex to deal with it more thoroughly.

We're aware of at least one apparent bug. In the presence of the
constraint `{ ... }` on type parameter `P`, as in `P: { ... }`, we
see a warning. Fair enough, it sounds like `P` shouldn't be inexact.
But that warning disappears when `P` is unconstrained! This smells
of a bug; when `P` is unconstrained, it may very well be an inexact
object, it's up to the caller. So we'd want it to still be flagged.

As Greg points out, a bug like this may be more forgivable because
this is just a lint rule, so it's allowed to be heuristic in some
cases.

See more discussion at
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Flow.20.60inexact-spread.60/near/886984.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.6