android sharing [nfc]: Flag duplicate code.
We're duplicating this code for now, for convenience at the
prototyping stage while we're still getting this feature working and
sorting out just how it should work.  But it's tricky code laden with
details about the Android framework and about RN, and so we won't want
to leave it this way.  Or if we neglect to fix that, we'll at least
want to have a warning for when modifying either copy of the logic.