lightbox: Use `tryGetFileDownloadUrl` to download images.
Using the helper `tryGetFileDownloadUrl`, we try to generate and
use temporary authless URLs to download files from the lighbox.
This fixes an issue where redirects to S3 caused RNFetchBlob to
fail downloading the file.

If we fail to download the file, which happens if we failed to
generate the temporary URL because the server does not support it,
we open it in the browser instead, from where the user can easily
save it.

Also makes changes in other places required due to this change.

Fixes: #2115