api [nfc]: Link `getFileTemporaryUrl` more specifically to our API spec.
The OpenAPI document that is our API spec is quite long, so it's
helpful to point to the specific place where this particular endpoint
is documented.

Then once we've found that, it seems best to match the local binding's
name to the operation ID in that spec, rather than let them differ by
just the order of a pair of words.  Considered in isolation our local
name might be the better one, but having names match across layers is
helpful when digging to understand just what some operation is doing.