flow: Use path-based imports instead of Haste.
Part of the RN v0.60 -> v0.61 changes to the template app [1],
corresponding to facebook/react-native@bf8d91868 and a correction in
facebook/react-native@732ded7f7. This must happen at or after the
main upgrade commit because the former declares that it depends on
facebook/react-native#25100, which landed in
facebook/react-native@69d1ed731 and was released in v0.61.0.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.60.6&to=0.61.5