ios deps: Remove `project.xcworkspace` from .gitignore.
Part of the RN v0.60 -> v0.61 changes to the template app [1],
corresponding to facebook/react-native@a7b9be4a1. This can freely be
done before the upgrade commit because the presence or absence of
this file isn't associated with the upgrade, but rather (it seems) a
detail of how the app was created in the very beginning.

Judging from the template app, it seems that this line is meant to
address the potential existence of a directory at
`ios/ZulipMobile.xcodeproj/project.xcworkspace`. We don't have a
directory there, whether or not it's git-ignored.

Greg mentions in discussion [2] a quote from a blog post [3] with
some plausible details:

"""
It is possible to just have an .xcodeproj visible to the
desktop, and have a .xcworkspace directory inside the .xcodeproj
directory. In this case, the workspace is always called
`project.xcworkspace` and is at the root directory of the project.
This is what you get if you use the Xcode wizard to create a new
project first, instead of creating a workspace and adding projects
to it.
"""

Greg concludes that if we did have a directory there, it would
belong in source control. So it doesn't belong in the .gitignore,
and it's good that they fixed that in the template.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.60.6&to=0.61.5
[2]: https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M3781.20RN.20v0.2E61.20upgrade/near/900957
[3]: http://neurocline.github.io/dev/2016/04/16/xcode-xcworkspace-and-xcodeproj.html