docs: Point to new home of RN setup instructions.
The page we've been pointing to now starts with a message saying:

  Welcome to the very start of your React Native journey! If you're
  looking for environment setup instructions, they've moved to their
  own section. Continue reading for an introduction to the
  documentation, Native Components, React, and more!

where "their own section" is a link.  For one of the places where we
link to this page, the material that's still on that page is the right
thing to link to; for the rest, update the links to save the reader an
indirection.