redux-persist-migrate types: Import `StoreEnhancer`.
With Flow v0.105.0, which we'll start using along with the RN v0.60
-> v0.61 upgrade [1], we get some new errors that this commit fixes.

We never precisely identified the cause of these errors [2], but it
also wasn't completely clear why we redefined `StoreEnhancer`
instead of importing it from Redux, in f3c6ff717. From reading the
commit message, it seems likely that we wanted to cut down on
complexity by not supporting some features that we don't use, which
makes sense as a strategy.

Kicking the tires, there don't appear to be useful checks being
suppressed.

[1]: https://react-native-community.github.io/upgrade-helper/?from=0.60.6&to=0.61.5
[2]: See discussion at https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Flow.20v0.2E105.2E0.20upgrade/near/899790.