fetchActions: Stop using `tryUntilSuccessful` in `fetchPrivateMessages`.
`tryUntilSuccessful` is about to grow a timeout, which we've had in
mind in particular for the initial fetch. It's also about to be
moved into this file, and we want to keep its association with
`doInitialFetch` by putting it directly above it, with nothing in
between.

Removing this call site lets us do so without ESLint complaining of
`no-use-before-define`. And Greg says it's all the same to him [1],
especially since `fetchPrivateMessages` is doomed to be deleted
anyway, as the JSDoc says.

[1]: https://github.com/zulip/zulip-mobile/issues/4165#issuecomment-645657326