crunchy shell: Use AvatarURL for User and CrossRealmBot objects.
For the `/register` response, we added a `transform` function with a
`migrateUserOrBot` helper in a recent commit, to be used here.

For events (for added and updated users), the natural place to do
this is in `eventToAction`. We don't reuse `migrateUserOrBot` from
`registerForEvents` because users and bots are likely to have
different shapes in the `/register` response and in events, and
we'll want to handle those shapes separately.