libdefs: Prepare to edit `redux-mock-store`.
In an upcoming commit, we'll change the Dispatch type in this libdef
to play nicely with our async redux-thunk actions. I don't see an
easy way to give the libdef the flexibility to give different types
depending on what middleware is used, so, we hard-code it for
ourselves, since we know we'll be using redux-thunk for the
foreseeable future.

This means we probably can't reasonably make a PR to FlowTyped, even
though this is (surprisingly) one of the few libraries we use that
does have a libdef hosted by FlowTyped. Ah, well.

Here, we just move the file, delete the metadata at the top, and run
our auto-formatting.

[cherry-picked from #4171]