ChatScreen: Fetch messages here, not `doInitialFetch` and `doNarrow`.
For more straightforward error handling, move the dispatch of
fetching messages to the UI component that we want to make
responsible for handling errors in the fetch (and that is already
responsible for showing the messages).

Now, we can wrap the fetch in a `try` in the UI component itself,
and easily set the local UI state in the `catch`. When we're in the
error state, we'll see an error view instead of the infinite loading
view, thus fixing #4186.

We have to carefully remove the fetches in `doInitialFetch` and
`doNarrow` that this strategy replaces, and edit the tests
accordingly. So, do.

We fetch in `componentDidMount`, but that's not quite enough: we
might need to reregister an event queue (see DEAD_QUEUE) while the
component is just sitting there, having been mounted for some time.
In particular, this needs to happen on REALM_INIT. This was handled
before by `doInitialFetch` being called upon DEAD_QUEUE, and the
fetch being done after dispatching REALM_INIT. So, now, we listen
for REALM_INIT as a state change in `state.session.eventQueueId` in
`componentDidUpdate`.

We also preserve that the fetch on REALM_INIT only happens for the
"topmost" or "last" narrows, using the same values that were used in
`doInitialFetch`. We do a few things better:

- Before, a separate fetch was done for the "topmost" *and* the
  "last" narrow, even if they were the same narrow. I haven't fully
  understood the differences between the two markers, but I suspect
  there's a lot of overlap. Now, a particular `ChatScreen` doesn't
  fetch twice, even if it's both "topmost" and "last" [1].

- Before, for non-"topmost" and non-"last" narrows, we didn't bother
  to fetch them at all, on reregistering an event queue -- even when
  they later became "topmost" or "last". So, set up a little bit of
  local UI state to say that we want them to fetch as soon as they
  next gain focus.

Finally, adjust tests and comments according to the changes. See
also discussion [2], where Greg carefully explains the subtle ways
actions have to be ordered, and we discuss some race conditions.

[1] N.B.: If narrow X is considered the "topmost" or "last" narrow,
    and multiple `ChatScreen`s are showing narrow X, multiple
    fetches will be done (one per `ChatScreen` showing X). Where the
    number of `ChatScreen`s showing X is large, this is a
    performance regression. This isn't likely to be an issue in
    practice, but Greg points out that it's a sign something's not
    quite modeled correctly, and we'll probably want to fix it:

    """
    I think specifically the issue is that the condition we'd like
    to have governing this isn't "is this ChatScreen showing the
    same narrow that's topmost or last", but rather "is this the
    topmost or last ChatScreen".
    """

    (from
    https://github.com/zulip/zulip-mobile/pull/4205#pullrequestreview-461268523)

[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4156.20Message.20List.20placeholders/near/948514

Fixes: #4186