Revert "android: Handle new intents in 'onNewIntent' when activity is resumed."
This reverts commit a9142412159d81af400a17d4ec18ff0c840cd347.

This definition overrode the onNewIntent implementation provided by
ReactActivity.  That one has an important job: in particular, on a
VIEW intent -- which is the kind that's used for opening a URL --
that onNewIntent implementation and its callees are responsible for
passing that information to the listener we've set in JS via the
Linking module.

And that, in turn, is what we use to complete the auth flow for
anyone logging into Zulip through OAuth or any other mechanism that
goes through the web.  So that functionality was broken by this
change.

In order to complete the share-from-another-app feature, we still
need to arrange to handle these intents when they come to an
existing activity -- but we'll need to be more careful with the
mechanisms provided by RN when doing so.

Discussion here:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/double.20initialization.20on.20sharing/near/950939