doEventActionSideEffects [nfc]: Turn into an action creator.
We're about to have this dispatch an action as a side effect.

This makes me think that the Redux Middleware API [1] is another
quite plausible place this stuff could go. But that might tempt a
too broad range of changes, such as those that have nothing to do
with server events.

(If we ran these side effects from Redux middleware, they would be
in a different "middle" than was meant by this function's previous
name, `eventMiddleware`. Redux middleware is run after you've put
together a well-formed action, not before.)

[1] https://redux.js.org/api/applymiddleware/