compose: Cut animation on mention-unsubscribed warning.
This animation looks kind of odd, and inappropriate for the entrance
of a banner.  See description here:
  https://github.com/zulip/zulip-mobile/pull/4101#issuecomment-669764675

Ideally we'd have an appropriate animation, like the banner sliding
in from the bottom.  But no animation at all actually looks pretty OK
to me here, and definitely better than this odd animation.

(We use the same odd animation for the "N unreads" banner, and it's
equally inappropriate there.  Likely we should just drop it there, too.)