react-intl libdef: Make some React type annotations useful.
Anything at `React.` is `any` [1]. So, in a few places, use Flow's
builtin React type annotations, like React$Node instead of
React.Node.

These are the obvious ones that we use. I tried exhaustively
translating everything from the `React.` namespace into a `React$`
equivalent, but things broke in mysterious ways. Better to do it
incrementally, then, and we can come back if we want more of it to
work.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/libdef.3A.20react-native-webview/near/896571