deps: Update `react-intl` from v3 to v5, the latest.
The upgrade guides for v4 [1] and v5 [2] are pretty short and
simple, so, might as well jump to the latest (currently 5.8.6) all
at once.

We restrict the version range to that single version (instead of
using a caret or a tilde), since `react-intl` doesn't invite its
users to expect breaking changes to be handled well [3]. We also
change our useless `typescript` dev-dependency's version to a range
that's in the intersection of the range `react-intl` gives for it,
and the range we found in 78daeb607.

With v4 (and continuing into v5), we get a big stroke of luck for
the libdef. There's a single TypeScript file that contains all the
necessary types [4], so we only need to run Flowgen once, on that
one file. This saves a lot of the kind of work we had to do for v3.

The v4 declared breaking changes don't apply to us.

One v5 breaking change applies to us:

- `FormattedMessage` render prop is no longer variadic.

We've only ever cared about the first argument passed to the render
prop, so, grab just that one, to minimally handle the API change. In
fact, it seems like a bug that we're not paying attention to any
others that are passed. But it looks like no others are being
passed, currently; we haven't explicitly made that happen. But we'll
stop using this confusing pattern in an upcoming commit; see
discussion [5].

[1] https://formatjs.io/docs/react-intl/upgrade-guide-4x

[2] https://formatjs.io/docs/react-intl/upgrade-guide-5x

[3] https://github.com/formatjs/formatjs/pull/2067#issuecomment-714171574

[4] except for React, which is normal, and which we'll address to
    some extent in the next commit.

[5] https://github.com/zulip/zulip-mobile/pull/4222#discussion_r468244576