deps: Redo ESLint upgrade to ^6.8.0, this time adding `core-js`.
A redo of 01593b3af, after we reverted it in e3e8d9a76.

One thing we realized [1] is that we don't have to wait for
`prettier-eslint` to fix the problem of its missing `core-js`
dependency; we can provide it ourselves. We do so by just adding it
directly to our own `dependencies`.

So, in addition to the commands listed in 01593b3af, also run
`yarn add core-js@^3.1.4`. We choose ^3.1.4 because that's what was
there before prettier/prettier-eslint#338. The open PR I mention in
e3e8d9a76 (prettier/prettier-eslint#359) suggests ^3.6.5, but it
seems to do so at random; might as well approach as closely as
possible a configuration that we know works.

Now, we have the ESLint upgrade, and format-on-save works as
expected.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/SOLVED.3A.20Prettier.20auto-formatting/near/982426

Fixes: #4120