typing status [nfc]: Encapsulate ensureTypingStatusExpiryLoop.
The condition about whether there are any existing typing-status
indicators is kind of mysterious when read literally, and even seems
backward: we only start trying to clear them if there are *none*?

The point of it is that that's also the break condition for the
loop that otherwise sits around continuously, looking for expirable
typing statuses and expiring them.  That's an implementation detail
of that loop, and kind of a subtle one.  Move it to sit next to the
loop's own implementation.