accounts: Make `Auth.realm` a URL object, including in `state.accounts`.
Changing the `Auth` type implicitly changes the `Account` and
`Identity` types, too.

Done with an effort to minimize unnecessary follow-on changes in
this commit. Here, we don't propagate the realm in its URL form as
far toward the eventual consumers of the realm as we'll eventually
want to; we'll do that in upcoming commits.

Wherever two realm values are compared for equality with `===`, we
make sure they are both stringified first. Flow didn't warn about
these cases, so we caught them in test failures and by scanning
through a full-project search for the term 'realm'.