webAuth: Better compare web auth callback's `realm` with current realm.
Part of #4146.

Rather than doing a straight string comparison, which will unfairly
reject on, e.g., a trailing slash in one URL but not the other,
parse them both with `react-native-url-polyfill`, made available at
the global URL in 413bf9521.

Also stop using `url-parse` at all; `authFromCallbackUrl` is the
only place where we use it. We introduced
`react-native-url-polyfill` in c675cdd63, for #4081, and we'll keep
using that for #4146. Not only is it specifically intended for RN
projects, but its README and implementation make it very clear that
it's an implementation of a standard; in particular, the WHATWG URL
Standard [1].

[1] https://url.spec.whatwg.org/