android: Use `Intl` variant of JavaScriptCore.
We don't want to be using JavaScriptCore forever, if we can avoid
it; we're tracking the switch to Hermes as #4131. In fact, Hermes
doesn't yet support Intl.

For now, this tiny code change will allow `react-intl` to work on
Android. A `react-intl` doc [1] links to instructions [2] [3] to use
a variant of JavaScriptCore that supports `Intl`. So, follow those
instructions.

Reportedly, this change will add 6MiB per architecture. #3547 was
recently resolved, though, giving us some wiggle room.

[1] https://formatjs.io/docs/react-intl/#react-native
[2] https://github.com/react-native-community/jsc-android-buildscripts#international-variant
[3] They're slightly out-of-date, currently, but it's still easy to
    tell what to do; see
    https://github.com/facebook/react-native/commit/d7f5153cd#diff-b8dd15c78827c1b48f9fe21c45686142R100-R111
    for the new edit to make to `android/app/build.gradle`.