jest: Make "modern" fake-timer implementation the default.
Also, remove several now-unnecessary calls of
`jest.useFakeTimers('modern')`, but keep a few assertions that the
"modern" timers are actually being used.

In particular, our `jestSetup` is a central place where we make the
assertion. Not only is it good to check that we still intentionally
set the "modern" implementation, but we want to make sure that the
setting is correctly applied. See the note in fb23341c3 about it
being silently not applied until we added @jest/source-map as a
direct dependency.

We have an ESLint rule, from 2faad06a8, preventing imports from
'**/__tests__/**'; the rule is active in all files not matching that
same pattern. Add an additional override so that we can make the
"modern"-timers assertion from within `jest/jestSetup.js`.