android build: Omit `entryFile`, now that we can.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@a0d874087. This must happen
at or after the main upgrade commit because the field is newly
optional at the new version.

There isn't a clear functional advantage to doing it this way. In
fact, the maintainer who merged it said he was "not a big fan" of
the change [2]. But he merged it anyway, on the grounds that it
makes the experience more consistent with iOS.

So, I suppose it does that for us, too, and in any case we're now
more consistent with the template app.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2
[2] https://github.com/facebook/react-native/pull/26769#pullrequestreview-301107655