android build [nfc]: Add a lint-suppression line.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@87f94d558. This may freely be
done before the upgrade commit.

This `// noinspection` line is mentioned in an obsolete Android
Studio doc on suppressing lint warnings [2]; it's not clear that it's
the best way to do what it says it does. But might as well follow
the template.

Greg says [3],

"""
Weird that that's no longer in the
[current docs](https://developer.android.com/studio/write/lint.html#config).

Looks like that only describes turning a rule off globally or for a
whole Java or Kotlin method, or XML element, at a time. A method can
be pretty long! It's good to be able to keep these ignores tightly
focused, so they don't accidentally hide unrelated new issues in the
future.

And it doesn't seem to describe any way at all that applies to a
part of a Groovy file (like all these *.gradle build scripts are.)
"""

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2
[2] http://tools.android.com/tips/lint/suppressing-lint-warnings (be
    sure to click "stop" before the timer runs out on the automatic
    redirect)
[3] https://github.com/zulip/zulip-mobile/pull/4244#discussion_r481793524